Change outline canvas kernel depending on if editing cell types or not
